+++ lib/filey-diff/data-sources/file_system.rb
@@ -3,8 +3,9 @@ require 'digest/md5'
 module Filey
   module DataSources
     class FileSystem < DataSource
-      def initialize(root_directory)
+      def initialize(root_directory, &block)
         @root_directory = root_directory
+        @when_filey_loaded = lambda { |filey| block.call filey } if block
       end
 
       private
@@ -17,7 +18,9 @@ module Filey
           name = file.scan(/.*\/(.*)/).first.first
           normalised_path = ".#{path}"
           md5 = Digest::MD5.hexdigest(File.read(file))
-          Filey.new(normalised_path, name, File.mtime(file), md5)
+          filey = Filey.new(normalised_path, name, File.mtime(file), md5)
+          @when_filey_loaded.call(filey) if @when_filey_loaded
+          filey
         }
       end
     end